==> 01_atom_substitution.ipynb <==
import pandas as pd
from tqdm.auto import tqdm
from ast import literal_eval
from pymatgen.ext.matproj import MPRester
from pymatgen.core import Composition

from htvs_module.htvs_mods import substitution, comp_to_form

mpr = MPRester('WgPaLVntnoS3X3tdlT')

# Li containing materials from Materials Project

data = mpr.query({"elements":'Li','nelements':{ "$gte" : 3 , "$lte" : 3}}, properties=['material_id',"pretty_formula",'elements','formula','cif'])
df = pd.DataFrame(data)
df.to_csv('./data/01_1_all_Li_material_from_MP.csv',index=False)

df = pd.read_csv('./data/01_1_all_Li_material_from_MP.csv')[:10]
df

for i in tqdm(range(len(df))): 
    formula_dict = literal_eval(df.iloc[i]['formula'])
    df.iloc[i]['formula'] = (list(formula_dict.keys())+list(formula_dict.values()))

## Selected element for structure generation

sub_el_list = ['Li','B','C','N','O','F','Na','Mg','Al','Si','P','S',
               'Cl','Ca','Zn','Ga','Ge','As','Se','Sr','Cd','In',
               'Sn','Te','I','Ba','Hg','Pb'] 

## Elemental substitution

substituted_list=substitution(df,sub_el_list)

subcomp_mpid_df = pd.DataFrame()
sub_comp_list = []
mpid_list = []
reduced_form_list = []
element = []
for i in tqdm(substituted_list):
    my_comp_list=i[:-1]
    mpid=i[-1]
    reduced_form=Composition(comp_to_form(my_comp_list)).reduced_formula
    
    sub_comp_list.append(my_comp_list)
    mpid_list.append(mpid)
    reduced_form_list.append(reduced_form)
subcomp_mpid_df['sub_comp'] = sub_comp_list
subcomp_mpid_df['mpid'] = mpid_list
subcomp_mpid_df['reduced_formula'] = reduced_form_list
subcomp_mpid_df

subcomp_mpid_df.to_csv('./data/01_2_substitution_final.csv',index=False)
subcomp_mpid_df = pd.read_csv('./data/01_2_substitution_final.csv')
subcomp_mpid_df
