==> 02_3_stability_window.ipynb <==
import pickle
import pandas as pd
import multiprocessing as multi
from tqdm.auto import tqdm
from joblib import Memory
from pymatgen.core import Composition

from htvs_module.htvs_mods import VirtualEntry, get_window_range

df = pd.read_csv('./data/03_3_after_band_gap_screening.csv')
df

# Get stability window

unique_formula_df = df[['pretty_formula','composition']]
unique_formula_df

num_cores = multi.cpu_count() # 8
comp_list = unique_formula_df['pretty_formula']

memory = Memory('./data/stability_multiprocesser',verbose=0)
calc_stability_cached = memory.cache(get_window_range)

with multi.Pool(num_cores) as pool:
    stability = list(tqdm(pool.imap(calc_stability_cached, comp_list),total = len(comp_list)))

stability_df = pd.DataFrame(stability,columns=['lower_bound','upper_bound','total_range'])
stability_df = pd.concat([unique_formula_df,stability_df],axis=1)
stability_df

stability_df.to_csv('./data/04_1_stability_window.csv',index=False)
stability_df=pd.read_csv('./data/04_1_stability_window.csv')

# Cathode voltage > 1V & Total stability window range > 0.001 (example)

screened_by_stability_window_df = stability_df[(stability_df['upper_bound']>1)&(stability_df['total_range']>0.001)]

# Stability window screening result

df

merged_result = pd.merge(screened_by_stability_window_df,df,on=['composition','pretty_formula'],how='inner')
merged_result.to_csv('./data/04_2_after_stability_window_screening.csv',index=False)

merged_result = pd.read_csv('./data/04_2_after_stability_window_screening.csv')
merged_result
