==> 03_1_get_structure.ipynb <==
import pandas as pd
from tqdm.auto import tqdm
from ast import literal_eval
from pymatgen.core.structure import Structure

from htvs_module.htvs_mods import structure_substitution

# Get base structures

ref_df = pd.read_csv('./data/01_1_all_Li_material_from_MP.csv',index_col='material_id')
for i in tqdm(range(len(ref_df))): 
    formula_dict = literal_eval(ref_df.iloc[i]['formula'])
    ref_df.iloc[i]['formula'] = (list(formula_dict.keys())+list(formula_dict.values()))
ref_df

# All substituted formula with MP id

all_df = pd.read_csv('./data/01_2_substitution_final.csv')
all_df['index'] = all_df.index
all_df

selected_df = pd.read_csv('./data/04_2_after_stability_window_screening.csv')
selected_df

# all screened formula with MP id

selected_structure_df = pd.merge(all_df, selected_df, left_on='reduced_formula',right_on='pretty_formula', how='inner')
selected_structure_df

selected_structure_df.to_csv('./data/05_1_before_get_structure.csv',index=False)
selected_structure_df=pd.read_csv('./data/05_1_before_get_structure.csv')
selected_structure_df

# Get substituted structures

changed_structure_list = []
for i, row in tqdm(selected_structure_df.iterrows(),total= len(selected_structure_df)):
    changed_structure_list.append(structure_substitution(row,ref_df))
len(changed_structure_list)

changed_structure_df = pd.DataFrame(changed_structure_list)
changed_structure_df.columns = ['index','sub_structure']
changed_structure_df

sub_structure_df = pd.merge(selected_structure_df,changed_structure_df,on='index',how='inner')
sub_structure_df

sub_structure_df.to_csv('./data/05_2_sub_structure_with_predictions.csv',index=False)
