==> 03_2_drop_duplicate.ipynb <==
import pandas as pd
from tqdm.auto import tqdm
from htvs_module.htvs_mods import get_indirect_structure

formula_and_structure_df = pd.read_csv('./data/05_2_sub_structure_with_predictions.csv')
formula_and_structure_df

# Get structrue representations and delete duplications

formula_and_structure_df = get_indirect_structure(formula_and_structure_df)
formula_and_structure_df

formula_and_structure_df.to_csv('./data/06_1_after_bulk_enumerator.csv',index=False)
formula_and_structure_df=pd.read_csv('./data/06_1_after_bulk_enumerator.csv')

drop_dupl_df = formula_and_structure_df.drop_duplicates(['reduced_formula','spacegroup','wyckoffs','species'],keep='first')
drop_dupl_df

drop_dupl_df.to_csv('./data/06_2_drop_duplicate.csv',index=False)
