==> 03_4_mxrd.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from tqdm.auto import tqdm
from io import StringIO
from ase.io import read

from pymatgen.io.ase import AseAtomsAdaptor
from pymatgen.analysis.diffraction import xrd
from pymatgen.core.spectrum import Spectrum

df = pd.read_csv('./data/06_2_drop_duplicate.csv')
df

# Get modified-XRD data and processing

mxrd = []
for ii in tqdm(range(len(df))):
    atoms = read(StringIO(df['sub_structure'][ii]),format='vasp')
    anions = ['O','F','S','Cl','Se','Br','I','Te'] # except 16, 17 group

    if len(set(atoms.get_chemical_symbols())) == 2:
        del atoms[[atom.index for atom in atoms if atom.symbol == 'Li']]
    else:
        del atoms[[atom.index for atom in atoms if not (atom.symbol in anions)]]

    for i in range(len(atoms)):
        atoms[i].symbol = 'S'
    v_offset = (atoms.get_global_number_of_atoms()*40)/atoms.cell.volume
    c_offset = (v_offset**(1/3))

    new_cell=atoms.cell*c_offset

    atoms.set_cell(new_cell,scale_atoms=True)

    struc = AseAtomsAdaptor.get_structure(atoms)
    pa=xrd.XRDCalculator(wavelength='CrKb1').get_pattern(structure=struc,scaled=True,two_theta_range=(0,89.98))

    threshold = 0
    psd_idxs = pa.y > threshold #array of 0 and 1
    pay = pa.y * psd_idxs #zero out all the unnecessary powers

    ra = [round(x,1) for x in np.arange(0,91,0.1)]
    # ra2 = [round(x,2) for x in np.arange(0,90.1,0.5)]
    ry = [0 for _ in (ra)]

    rdf = pd.DataFrame(columns=ra)
    rdf.loc[0]=(ry)
    rdf

    rrdf = rdf[:]
    rx = [round(i,1) for i in pa.x]
    for i in range(len(rx)):    
        s=rx[i]
        rrdf[s] = max(rrdf[s][0],pay[i])

    spec = Spectrum(ra[:900],list(rrdf.loc[0])[:900])


    ssp = spec
    ssp.smear(0.2,'gaussian')
    ssp.normalize('sum',10)
    mxrd.append(list(ssp.y))

df['mxrd']=mxrd
df.to_csv('./data/08_1_bulk_mxrd.csv',index=False)

df_new_only_mxrd = pd.read_csv('./data/08_1_bulk_mxrd.csv')[['composition','mxrd','mpid']]
df_ref = pd.read_csv('./data/07_3_mxrd_from_ref.csv')[['formula_id','r_xrd','cond']]
df_ref.columns = ['composition','mxrd','conductivity']

df_new_only_mxrd

df_new_and_old = pd.concat([df_new_only_mxrd,df_ref]).reset_index()
df_new_and_old.drop(['index'],axis=1,inplace=True)
df_new_and_old

df_new_and_old.to_csv('./data/08_2_bulk_mxrd_merged.csv',index=False)
